Extract step cost-threshold logic and slim descendant queries
